+++ src/mainGkmKernel.cpp
@@ -28,51 +28,25 @@
 
 #include "global.h"
 #include "globalvar.h"
+#include "gkmOptions.h"
 
 #include "Sequence.h"
-#include "CountKLmers.h"
-#include "CountKLmersGeneral.h"
-#include "CountKLmersH.h"
 #include "CalcWmML.h"
-#include "MLEstimKLmers.h"
-#include "MLEstimKLmersLog.h"
-#include "KLmer.h"
 #include "SequenceNames.h"
-#include "EstimLogRatio.h"
 #include "LTree.h"
 #include "LTreef.h"
 #include "LTreeS.h"
 #include "LList.h"
 #include "CiDLPasses.h"
-#include "GTree2.h"
 
 using namespace std;
 
 //#include "stdafx.h"
 
-typedef struct {
-	int L;
-	int K;
-	int maxnmm;
-	int maxseqlen;
-	int maxnumseq;
-	int useTgkm;
-	bool addRC;
-	bool usePseudocnt;
-	bool OutputBinary;
-	char *posfile;
-	char *negfile;
-	char *outfile;
-    double wildcardLambda;  //parameter lambda for (LK2004)
-    int wildcardMismatchM;  //parameter M for wildcard or mismatch kernels (LK2004)
-    char *alphabetFN; //alphabets file name
-    int maxnThread; // max number of threads
-} OptsGkmKernel;
-
 int gkmKernelSimple(OptsGkmKernel &opt);
 int gkmKernelSuffixTree(OptsGkmKernel &opt);
 
-void print_usage_and_exit_gkmKernel(char *prog)
+void print_usage_and_exit_gkmKernel(const char *prog)
 {
     Printf("\n");
     sprintf(globtmpstr, " Usage: %s [options] <pos_seqfile> <neg_seqfile> <outfile>\n",prog );Printf(globtmpstr);
@@ -100,130 +74,94 @@ void print_usage_and_exit_gkmKernel(char *prog)
 	sprintf(globtmpstr,"%s", "                 default=0\n"); Printf(globtmpstr);
 	sprintf(globtmpstr,"%s", "  -R             if set, reverse complement sequences will NOT be considered\n"); Printf(globtmpstr);
 	sprintf(globtmpstr,"%s", "  -p             if set, a constant to will be added to the count estimates\n"); Printf(globtmpstr);
-	//cout << "  -b             if set, the output matrix will be stored in a binary format" << endl;
 	sprintf(globtmpstr,"%s", "  -M             max mismatch for Mismatch kernel or wildcard kernel, default=2\n"); Printf(globtmpstr);
 	sprintf(globtmpstr,"%s", "  -L             lambda for wildcard kernel, defaul=1.0\n"); Printf(globtmpstr);
     sprintf(globtmpstr,"%s", "  -A             alphabets file name, if not specified, it is assumed the inputs are DNA sequences \n"); Printf(globtmpstr);
     sprintf(globtmpstr,"%s", "  -T             maximum number of threads, defaul=2*l\n"); Printf(globtmpstr);
     
     Printf(" \n");
-
-	//exit(0);
 }
-int mainGkmKernel(int argc, char** argv) //mainGkmKernel
+
+// Parse argv into opt. Returns 0 on success, 1 if usage was printed (nothing to run).
+static int gkmKernelParseArgs(int argc, char** argv, OptsGkmKernel &opt)
 {
-    /*printf("welcome!\n");
-    printf("\n argc = %d \n", argc);
-    for(int i=0;i<argc;i++){
-        printf("%s\n",argv[i]);
-    }*/
-    
-
-	OptsGkmKernel opt;
-    
-#ifdef __GLIBC__
-    ::optind = 0; // glibc re-initialises getopt only when optind is 0; optind = 1 leaves its internal state from the previous call
-#else
-    ::optind = 1; ::optreset = 1; // BSD / macOS
-#endif
-
+    ::optind=1; // reset getopt()
     int c;
-
-	opt.L = DEF_L; 
-	opt.K = DEF_K; 
-	opt.maxnmm = DEF_D;
-	opt.maxseqlen = DEF_MAXSEQLEN; 
-	opt.maxnumseq = DEF_MAXNUMSEQ; 
-	opt.useTgkm = DEF_TGKM; 
-	opt.addRC = true; 
-	opt.usePseudocnt=false; 
-	opt.OutputBinary=false; 
-    
-    opt.wildcardMismatchM= 2;
-    opt.wildcardLambda = 1.0;
-    opt.alphabetFN = NULL;
-    opt.maxnThread = 1000;
-    
-	int alg = 0;
-    
-    if (argc == 1) { print_usage_and_exit_gkmKernel(argv[0]); return 0;}
+    if (argc == 1) { print_usage_and_exit_gkmKernel(argv[0]); return 1;}
 
 	while ((c = getopt (argc, argv, "l:k:d:m:n:t:a:L:M:A:T:Rpb")) != -1)
 	{
 		switch (c) 
 		{
-			case 'l':
-                //printf("\ngetopt  l = %s = %d\n",optarg,atoi(optarg)  );
-				opt.L = atoi(optarg);
-				break;
-			case 'k':
-				opt.K = atoi(optarg);
-				break;
-			case 'd':
-				opt.maxnmm = atoi(optarg);
-				break;
-			case 'm':
-				opt.maxseqlen = atoi(optarg);
-				break;
-			case 'n':
-				opt.maxnumseq = atoi(optarg);
-				break;
-			case 't':
-				opt.useTgkm = atoi(optarg);
-				break;
-			case 'a':
-				alg = atoi(optarg);
-				break;
-			case 'R':
-				opt.addRC = false;
-				break;
-			case 'p':
-				opt.usePseudocnt = true;
-				break;
-			case 'b':
-				opt.OutputBinary = true;
-				break;
-			case 'M':
-				opt.wildcardMismatchM = atoi(optarg);
-				break;
-			case 'L':
-				opt.wildcardLambda = atof(optarg);
-				break;
-			case 'A':
-				opt.alphabetFN = optarg;
-				break;
-  		case 'T':
-  		  opt.maxnThread = atoi(optarg);
-  		  break;
-  		default:
-                print_usage_and_exit_gkmKernel(argv[0]);return 0;
+			case 'l': opt.L = atoi(optarg); break;
+			case 'k': opt.K = atoi(optarg); break;
+			case 'd': opt.maxnmm = atoi(optarg); break;
+			case 'm': opt.maxseqlen = atoi(optarg); break;
+			case 'n': opt.maxnumseq = atoi(optarg); break;
+			case 't': opt.useTgkm = atoi(optarg); break;
+			case 'a': opt.alg = atoi(optarg); break;
+			case 'R': opt.addRC = false; break;
+			case 'p': opt.usePseudocnt = true; break;
+			case 'b': opt.OutputBinary = true; break;
+			case 'M': opt.wildcardMismatchM = atoi(optarg); break;
+			case 'L': opt.wildcardLambda = atof(optarg); break;
+			case 'A': opt.alphabetFN = optarg; break;
+			case 'T': opt.maxnThread = atoi(optarg); break;
+			default: print_usage_and_exit_gkmKernel(argv[0]); return 1;
 		}
 	}
 
-	if (argc-optind != 3) { print_usage_and_exit_gkmKernel(argv[0]); return 0;}
+	if (argc-optind != 3) { print_usage_and_exit_gkmKernel(argv[0]); return 1;}
 
 	int index = optind;
 	opt.posfile = argv[index++];
 	opt.negfile = argv[index++];
 	opt.outfile = argv[index++];
+	return 0;
+}
 
+int mainGkmKernel(int argc, char** argv) //mainGkmKernel
+{
+	OptsGkmKernel opt;
+	if (gkmKernelParseArgs(argc, argv, opt) != 0) return 1;
+	return gkmKernelRun(opt);
+}
+
+// Validate, load the alphabet, dispatch. Shared by the CLI and the R binding.
+int gkmKernelRun(OptsGkmKernel &opt)
+{
 	//check parameters
+	if (opt.L < 1 || opt.K < 1)
+	{
+		Printf("\n ERROR: L and K must be positive.\n"); return 1;
+	}
 	if ((opt.K > opt.L) &&(opt.useTgkm<3))
 	{
-		Printf("K must be less than or equal to L!\n\n");
-        print_usage_and_exit_gkmKernel(argv[0]); return 0;
+		Printf("\n ERROR: K must be less than or equal to L!\n"); return 1;
 	}
 
 	if ((opt.maxnmm > 0) && (opt.L < opt.maxnmm))
 	{
-		Printf("maxMismatch must be less than or equal to L!\n\n");
-        print_usage_and_exit_gkmKernel(argv[0]); return 0;
+		Printf("\n ERROR: maxMismatch must be less than or equal to L!\n"); return 1;
+	}
+	if (opt.useTgkm < 0 || opt.useTgkm > 4)
+	{
+		Printf("\n ERROR: filter type (-t) must be between 0 and 4.\n"); return 1;
+	}
+	if (opt.alg < 0 || opt.alg > 2)
+	{
+		Printf("\n ERROR: algorithm (-a) must be 0, 1 or 2.\n"); return 1;
+	}
+	if (opt.maxseqlen < opt.L || opt.maxnumseq < 1)
+	{
+		Printf("\n ERROR: maxSeqLen must be at least L and maxNumSeq at least 1.\n"); return 1;
 	}
 
 	// the alphabet lives in a global that outlives this call: start every call from DNA
 	globalConverter.resetToDNA();
-	if (opt.alphabetFN!=NULL){
-		if (globalConverter.readAlphabetFile(opt.alphabetFN, MAX_ALPHABET_SIZE)!=0) return 1;
+	int alg = opt.alg;
+	if (!opt.alphabetFN.empty()){
+		if (globalConverter.readAlphabetFile(opt.alphabetFN.c_str(), MAX_ALPHABET_SIZE)!=0) return 1;
 		if (opt.addRC&&(globalConverter.b!=4)&&(::globalConverter.b!=16)){
 			opt.addRC=false;
 			Printf("\nAdd Reverse Complement option is turned off.\n");
@@ -247,10 +185,8 @@ int mainGkmKernel(int argc, char** argv) //mainGkmKernel
 			}
 		case 1:
 			return gkmKernelSimple(opt);
-		case 2:
-			return gkmKernelSuffixTree(opt);
 		default:
-            print_usage_and_exit_gkmKernel(argv[0]); return 0;
+			return gkmKernelSuffixTree(opt);
 	}
 }
 
@@ -279,9 +215,9 @@ int gkmKernelSimple(OptsGkmKernel &opt)  //Use XOR precomputed hash table
 	bool addRC = opt.addRC;
 	//bool usePseudocnt= opt.usePseudocnt; 
 
-	char *posSeqsFN = opt.posfile;
-	char *negSeqsFN = opt.negfile;
-	char *outFN = opt.outfile;
+	const char *posSeqsFN = opt.posfile.c_str();
+	const char *negSeqsFN = opt.negfile.c_str();
+	const char *outFN = opt.outfile.c_str();
 
 	CLList **seqsL = new CLList *[nMAXSEQUENCES];
 	double *norm = new double [nMAXSEQUENCES];
@@ -449,7 +385,7 @@ int gkmKernelSimple(OptsGkmKernel &opt)  //Use XOR precomputed hash table
 		{
 			if(i>j)
 			{
-				fprintf(fo, "%e\t",(norm[i]*norm[j]<1E-50)?0.0:seqsL[i]->calcInnerProd(seqsL[j],c,mmcnt)/(norm[i]*norm[j]));
+				fprintf(fo, "%e\t",gkmCanon((norm[i]*norm[j]<1E-50)?0.0:seqsL[i]->calcInnerProd(seqsL[j],c,mmcnt)/(norm[i]*norm[j])));
 			}
 			else if (i==j)
 			{
@@ -549,9 +485,9 @@ int gkmKernelSuffixTree(OptsGkmKernel &opt)  //maingKernel
   bool addRC = opt.addRC;
   bool usePseudocnt= opt.usePseudocnt; 
   
-  char *posSeqsFN = opt.posfile;
-  char *negSeqsFN = opt.negfile;
-  char *outFN = opt.outfile;
+  const char *posSeqsFN = opt.posfile.c_str();
+  const char *negSeqsFN = opt.negfile.c_str();
+  const char *outFN = opt.outfile.c_str();
   
   int i = 1; 
   
@@ -631,20 +567,8 @@ int gkmKernelSuffixTree(OptsGkmKernel &opt)  //maingKernel
   int nneg=0;
   
   gMAXMM=maxnmm; //MaxMismatch
-  int UseGTree = 0; // GTree algorithm for speed
-  //if(UseGTree){
-  //int maxn=(1<<(2*L)) * ::Combinations(L, gMAXMM);
-  // if(maxn>100000000){maxn=100000000;};
-  // int maxn=100000000;
-  // gGTreeLeaves=new GTreeLeafData[maxn]; // list of all the leaf nodes
-  // gGTreeLeavesCnt=0; // number of all leaf nodes
-  
-  //}
-  
   
   CLTreeS *seqsTS= new CLTreeS();
-  //    GTree *seqsGTree= new GTree();
-  //   GTree2 *seqsGTree2= new GTree2();
   
   
   
@@ -806,24 +730,7 @@ int gkmKernelSuffixTree(OptsGkmKernel &opt)  //maingKernel
   
   int uniqueLmerCnt = seqsTS->leavesCount(0,L, globalConverter.b, nodesAtDepthCnt);
   sprintf(globtmpstr,"\n npos %d \n nneg %d \n  ntotal %d \n nunique %d\n",npos,nneg,ntotal,uniqueLmerCnt);Printf(globtmpstr);
-  int minL2 = L; if (minL2<2) minL2 = 2; 
-#ifdef USE_GLOBAL
-  for(int i=0;i<=minL2;i++)
   {
-    //gDFSlist[i] = new LTreeSnodeData *[uniqueLmerCnt];
-    //	gDFSlistT[i] = new CLTreeSptr *[uniqueLmerCnt];  // without nonEmptyDaughterCnt
-    gDFSlistT[i] = new CLTreeS *[uniqueLmerCnt];  // with nonEmptyDaughterCnt
-    gDFSMMlist[i] = new int[uniqueLmerCnt]; 
-    gDFSMMtree[i] = new CbinMMtree *[uniqueLmerCnt];
-  }
-  //int *curmmcnt = gDFSMMlist[0];
-  
-  //gDFSlistT[0][0] = seqsTS->daughter; // without nonEmptyDaughterCnt
-  gDFSlistT[0][0] = seqsTS; // with nonEmptyDaughterCnt
-  gDFSMMlist[0][0] = 0; 
-#endif
-  //   int UseGTree = 1;
-  if (!UseGTree){
     // if no IDL bound
     /*
     seqsTS->DFST(gDFSlistT[0],1, gDFSMMlist[0], 0, globalConverter.b);
@@ -921,74 +828,10 @@ int gkmKernelSuffixTree(OptsGkmKernel &opt)  //maingKernel
       delete []myThreads;
     }
     
-  }else{
-    gGTreeLeaves2=new GTreeLeafData2[uniqueLmerCnt * ::Combinations(L, gMAXMM)]; // list of all the leaf nodes
-    gGTreeLeavesCnt=0; // number of all leaf nodes
-    GTree2 *seqsGTree2= new GTree2();
-    int *tmpArray1 = new int[L];
-    
-    seqsTS->addToGTree(seqsGTree2, L,tmpArray1, MAX_ALPHABET_SIZE, L);
-    delete []tmpArray1;
-    sprintf(globtmpstr," gGTreeLeavesCnt = %d \n",gGTreeLeavesCnt);Printf(globtmpstr);
-    
-    for(int i=0;i<gGTreeLeavesCnt;i++){
-      gGTreeLeaves2[i].process();
-    }
-    // normalize
-    
-    for(int i=0;i<nseqs;i++)
-    {
-      
-      //            gMMProfile[i][0][i]/=Combinations(L, gMAXMM);;
-      
-      for (int j=0;j<=gMAXMM;j++)
-      {
-        //              int s=::Combinations(L-j, L-gMAXMM);
-        for(int k=0;k<=i;k++)
-        {
-          int mmijk = (gMMProfile[i][j][k]+gMMProfile[k][j][i]);
-          //                mmijk=mmijk/s;
-          gMMProfile[i][j][k]=gMMProfile[k][j][i]=mmijk;
-        }
-      }
-    }
-    
-    for(int i=0;i<nseqs;i++)
-    {
-      
-      //            gMMProfile[i][0][i]/=Combinations(L, gMAXMM);;
-      
-      for (int j=0;j<=gMAXMM;j++)
-      {
-        int s=::Combinations(L-j, L-gMAXMM);
-        //                int s=::Combinations(gMAXMM, j);
-        for(int k=0;k<nseqs;k++)
-        {
-          //                    gMMProfile[i][j][k]*=2;
-          int z=gMMProfile[i][j][k]/s;gMMProfile[i][j][k]=z;
-        }
-      }
-    }
-    
-    
-    for(int i=0;i<nseqs;i++)
-    {
-      
-      gMMProfile[i][0][i]+=LmersCnt[i];//self;
-    }
   }
   
   delete []nodesAtDepthCnt;
   
-  //
-#ifdef USE_GLOBAL
-  for(int i=0;i<=minL2;i++)
-  {
-    delete []gDFSlistT[i];
-    delete []gDFSMMlist[i];
-    delete []gDFSMMtree[i];
-  }
-#endif
   /*
   for(int i=0;i<nseqs;i++)
   {
@@ -1085,11 +928,11 @@ int gkmKernelSuffixTree(OptsGkmKernel &opt)  //maingKernel
         {
           if (usePseudocnt)
           {
-            fprintf(fo, "%e\t",(norm[i]*norm[j]<1E-50)?0.0:calcinnerprod(i,j,c, n0,C,LmersCnt[i], LmersCnt[j], btL)/(norm[i]*norm[j]));
+            fprintf(fo, "%e\t",gkmCanon((norm[i]*norm[j]<1E-50)?0.0:calcinnerprod(i,j,c, n0,C,LmersCnt[i], LmersCnt[j], btL)/(norm[i]*norm[j])));
           }
           else
           {
-            fprintf(fo, "%e\t",(norm[i]*norm[j]<1E-50)?0.0:calcinnerprod(i,j,c)/(norm[i]*norm[j]));
+            fprintf(fo, "%e\t",gkmCanon((norm[i]*norm[j]<1E-50)?0.0:calcinnerprod(i,j,c)/(norm[i]*norm[j])));
           }
           
         }